+++ src/GAP.jl
@@ -261,6 +261,11 @@ function __init__()
       Globals.SetUserPreference(GapObj("utils"), GapObj("DownloadMaxTime"), 30)
     end
 
+    # Load the JuliaInterface manual (from a relative path of the juliainterface.so file)
+    juliaInterface_pkginfo = only(GAP.Globals.PackageInfo(GapObj("juliainterface")))
+    juliaInterface_pkgdoc = only(juliaInterface_pkginfo.PackageDoc)
+    GAP.Globals.HELP_ADD_BOOK(juliaInterface_pkgdoc.BookName, juliaInterface_pkgdoc.LongTitle, GAP.Globals.Directory(GapObj(abspath(dirname(JuliaInterface_path), "..", "..", "share", "doc", "JuliaInterface", "doc"))))
+
     if !isdefined(Main, :__GAP_ARGS__)
         # We had started GAP with the `-b` option.
         # Reset this option in order to leave it to GAP's `LoadPackage`