+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 - Set a timeout of 30 seconds for GAP's `Download` function,
   except if the user's `gap.ini` file defines already a finite timeout value.
+- Make the manual of the `JuliaInterface` GAP package available for interactive
+  access from a GAP session.
 
 ## Version 0.16.6 (released 2026-04-30)